+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidc/OidcSiopCombinedProtocolTest.kt
@@ -268,10 +268,7 @@ private suspend fun Holder.storeJwtCredential(
     credentialScheme: ConstantIndex.CredentialScheme,
 ) {
     storeCredential(
-        IssuerAgent(
-            EphemeralKeyWithoutCert(),
-            DummyCredentialDataProvider(),
-        ).issueCredential(
+        IssuerAgent().issueCredential(
             DummyCredentialDataProvider().getCredential(
                 holderKeyMaterial.publicKey,
                 credentialScheme,
@@ -286,10 +283,7 @@ private suspend fun Holder.storeSdJwtCredential(
     credentialScheme: ConstantIndex.CredentialScheme,
 ) {
     storeCredential(
-        IssuerAgent(
-            EphemeralKeyWithoutCert(),
-            DummyCredentialDataProvider(),
-        ).issueCredential(
+        IssuerAgent().issueCredential(
             DummyCredentialDataProvider().getCredential(
                 holderKeyMaterial.publicKey,
                 credentialScheme,
@@ -303,10 +297,7 @@ private suspend fun Holder.storeIsoCredential(
     holderKeyMaterial: KeyMaterial,
     credentialScheme: ConstantIndex.CredentialScheme,
 ) = storeCredential(
-    IssuerAgent(
-        EphemeralKeyWithSelfSignedCert(),
-        DummyCredentialDataProvider(),
-    ).issueCredential(
+    IssuerAgent(EphemeralKeyWithSelfSignedCert()).issueCredential(
         DummyCredentialDataProvider().getCredential(
             holderKeyMaterial.publicKey,
             credentialScheme,