+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Release 5.1.0:
  - Add `credentialScheme` and `subjectPublicKey` to internal `CredentialToBeIssued`
  - Refactor `issueCredential` of `Issuer` to directly get the credential-to-be-issued
  - Replace `buildIssuerCredentialDataProviderOverride` in `CredentialIssuer` with `credentialProvider` to extract user information into a credential
+ - Remove `dataProvider` from `IssuerAgent`s constructor, as it is not needed with the new issuing interface anyway
 
 Release 5.0.1:
  - Update JsonPath4K to 2.4.0